addressing ctd_plots
