+++ ctdcal/ctd_plots.py
@@ -1,293 +1,5 @@
 """
 A module for plotting and visualizing the CTDCAL data products.
 """
-import logging
-from pathlib import Path
 
-import gsw
-import matplotlib.pyplot as plt
-import matplotlib.ticker as ticker
-import numpy as np
-import pandas as pd
-
-log = logging.getLogger(__name__)
-
-
-def _apply_default_fmt(xlim, ylim, xlabel, ylabel, title, grid, fontsize=12):
-    plt.xlim(xlim)
-    plt.xticks(rotation=45)
-    plt.ylim(ylim)
-    plt.xlabel(xlabel, fontsize=fontsize)
-    plt.ylabel(ylabel, fontsize=fontsize)
-    plt.title(title, fontsize=fontsize)
-    plt.grid(grid)
-    plt.tight_layout()
-
-
-def _save_fig(ax, f_out):
-    if f_out is not None:
-        f_parent = Path(f_out).parent
-        if not f_parent.exists():
-            log.info(f"Parent folder '{f_parent.as_posix()}' doesn't exist... creating")
-            Path(f_out).parent.mkdir(parents=True)
-        plt.savefig(f_out)
-        plt.close()
-    else:
-        return ax
-
-
-def residual_vs_pressure(
-    param,
-    ref,
-    prs,
-    stn=None,
-    xlim=(-0.02, 0.02),
-    ylim=(6000, 0),
-    xlabel="Residual",
-    ylabel="Pressure (dbar)",
-    auto_title=True,
-    grid=False,
-    deep=False,
-    f_out=None,
-):
-    """
-    Plot residuals (ref - param) as a function of pressure.
-
-    Parameters
-    ----------
-    param : pd.Series or array-like
-        Input variable
-    ref : pd.Series or array-like
-        Reference variable
-    prs : pd.Series or array-like
-        Pressure variable
-    stn : pd.Series or array-like, optional
-        Station variable
-    xlim : tuple, optional
-        Lower and upper x-limits
-    ylim : tuple, optional
-        Lower and upper y-limits
-    xlabel : str, optional
-        Label for the x-axis
-    ylabel : str, optional
-        Label for the y-axis
-    auto_title : bool, optional
-        Generate title from input attributes (iff dtype is pd.Series)
-    grid : bool, optional
-        Overlay grid on figure
-    deep : bool, optional
-        Whether to plot only values >2000 dbar
-    f_out : path-like, optional
-        Path to save figure
-
-    Returns
-    -------
-    ax : matplotlib.axes, optional
-        Formatted scatter plot (if f_out is not given)
-    """
-    diff = ref - param
-    deep_text = " (>2000 dbar)" if deep else ""
-    if deep:
-        deep_rows = prs > 2000
-        diff, prs, stn = diff[deep_rows], prs[deep_rows], stn[deep_rows]
-
-    # initialize figure
-    plt.figure(figsize=(7, 6))
-    ax = plt.axes()
-
-    # color scatter by stations if given
-    if stn is not None:
-        idx, uniques = pd.factorize(stn)  # find unique stations #s and index them
-        sc = ax.scatter(diff, prs, c=idx, marker="+")
-        cbar = plt.colorbar(sc, ax=ax, pad=0.1)  # set cbar ticks to station names
-        tick_inds = cbar.get_ticks().astype(int)
-        cbar.ax.yaxis.set_major_locator(ticker.FixedLocator(tick_inds))
-        cbar.ax.set_yticklabels(uniques[tick_inds])
-        cbar.ax.set_title("Station")
-    else:
-        sc = ax.scatter(diff, prs, marker="+")
-
-    # formatting
-    title = None
-    if auto_title:
-        try:
-            title = f"{ref.name}-{param.name}{deep_text} vs. {prs.name}"
-        except AttributeError:
-            log.warning(
-                "Failed to set title from variable names (requires dtype pd.Series)"
-            )
-            log.info("Set afterward using 'ax.set_title(\"title\")'")
-    _apply_default_fmt(xlim, ylim, xlabel, ylabel, title, grid)
-
-    # save to path or return axis
-    return _save_fig(ax, f_out)
-
-
-def residual_vs_station(
-    param,
-    ref,
-    prs,
-    stn,
-    ylim=(-0.02, 0.02),
-    xlabel="Station Number",
-    ylabel="Residual",
-    grid=False,
-    deep=False,
-    f_out=None,
-):
-    """
-    Plot residuals (ref - param) as a function of station number.
-
-    Parameters
-    ----------
-    param : pd.Series or array-like
-        Input variable
-    ref : pd.Series or array-like
-        Reference variable
-    prs : pd.Series or array-like
-        Pressure variable
-    stn : pd.Series or array-like, optional
-        Station variable
-    ylim : tuple, optional
-        Lower and upper y-limits
-    xlabel : str, optional
-        Label for the x-axis
-    ylabel : str, optional
-        Label for the y-axis
-    grid : bool, optional
-        Overlay grid on figure
-    deep : bool, optional
-        Whether to plot only values >2000 dbar
-    f_out : path-like, optional
-        Path to save figure
-
-    Returns
-    -------
-    ax : matplotlib.axes, optional
-        Formatted scatter plot (if f_out is not given)
-    """
-    diff = ref - param
-    deep_text = " (>2000 dbar)" if deep else ""
-    if deep:
-        deep_rows = prs > 2000
-        diff, prs, stn = diff[deep_rows], prs[deep_rows], stn[deep_rows]
-
-    plt.figure(figsize=(7, 6))
-    ax = plt.axes()
-    sc = ax.scatter(stn, diff, c=prs, marker="+")
-    cbar = plt.colorbar(sc, pad=0.1)
-    cbar.set_label("Pressure (dbar)")
-
-    # formatting
-    try:
-        title = f"{ref.name}-{param.name}{deep_text} vs. {stn.name}"
-    except AttributeError:
-        title = None
-        log.warning(
-            "Failed to set title from variable names (requires dtype pd.Series)"
-        )
-        log.info('Set afterward using \'ax.set_title("title")`')
-    _apply_default_fmt(None, ylim, xlabel, ylabel, title, grid)
-
-    # save to path or return axis
-    return _save_fig(ax, f_out)
-
-
-def _intermediate_residual_plot(
-    diff,
-    prs,
-    stn,
-    xlim=(-0.02, 0.02),
-    xlabel="Residual",
-    show_thresh=False,
-    f_out=None,
-):
-    """
-    Internal function to make figures at intermediate processing stages for debugging.
-    """
-    ax = residual_vs_pressure(
-        0, diff, prs, stn=stn, xlim=xlim, xlabel=xlabel, auto_title=False, grid=True
-    )
-
-    if show_thresh:
-        thresh = np.array([0.002, 0.005, 0.010, 0.020])
-        p_range = np.array([6000, 2000, 1000, 500])
-        thresh = np.append(thresh, thresh[-1])  # this should still work fine even when
-        p_range = np.append(p_range, 0)  # thresh/p_range are defined elsewhere
-        plt.step(thresh, p_range, ":k")
-        plt.step(-thresh, p_range, ":k")
-
-    mean = np.round(np.nanmean(diff), 4)
-    stdev = np.round(np.nanstd(diff), 4)
-    ax.set_title(f"Mean: {mean} / Stdev: {stdev}")
-
-    # save to path or return axis (primarily for testing)
-    return _save_fig(ax, f_out)
-
-
-def param_vs_param(param1, label1, param2, label2, f_out=None, stn=None, tsT=None, tsS=None):
-    """
-    Scatter plots two CTD parameters against each other with support for TS plotting.
-
-    Support for bottle or time data.
-
-    Most TS plots have temperature on the Y axis and salinity on the X axis.
-    
-    Parameters
-    ----------
-    param1 : array-like of float
-        The first column/array for plotting on the X axis
-    label1 : string
-        Label for the X axis, attributed to param1
-    param2 : array-like of float
-        The second column/array for plotting on the Y-axis
-    label2 : string
-        Label for the Y axis, attributed to param2
-    f_out : String, optional
-        Path and filename to save TS plot
-    stn : array-like, optional
-        An array of station numbers or labels for 
-    tsT : array-like of float, optional
-        Temperature array for use in density gradient calculations
-    tsS : array-like of float, optional
-        Salinity array for use in density gradient calculations
-    """
-
-    fig = plt.figure(figsize=(7, 6))
-    ax = fig.add_subplot()
-
-    if stn is not None:
-        stn = np.array(stn)  # Ensure stn is a NumPy array
-        sc = ax.scatter(param1, param2, c=stn.astype(int), marker="+")
-        sc.set_clim(vmin=stn.astype(int).min(), vmax=stn.astype(int).max()) #   Color mapping based on station number
-        cbar = plt.colorbar(sc, format=ticker.FormatStrFormatter("%.0f"))
-        cbar.set_label("Station/Cast Label")
-    else:
-        sc = ax.scatter(param1, param2, color='black', marker="+")
-
-    ax.set_xlabel(label1)
-    ax.set_ylabel(label2)
-
-    #   TS plotting - define density contours
-    if tsT is not None:
-        #   Demarcate temperature, salinty, and limits for density
-        smin = tsS.min() - (0.01 * tsS.min())
-        smax = tsS.max() + (0.01 * tsS.max())
-        tmin = tsT.min() - (0.1 * tsT.max())
-        tmax = tsT.max() + (0.1 * tsT.max())
-        #   Gridding for a contour
-        xdim = round((smax - smin) / 0.1 + 1.0)
-        ydim = round((tmax - tmin) + 1.0)
-        #   Creating density vectors to fill the grid with densities
-        dens = np.zeros((ydim, xdim))
-        si = np.linspace(1, xdim - 1, xdim) * 0.1 + smin
-        ti = np.linspace(1, ydim - 1, ydim) + tmin
-        for j in range(0, int(ydim)):
-            for i in range(0, int(xdim)):
-                dens[j, i] = gsw.rho(si[i], ti[j], 0)
-        dens = dens - 1000  #   Convert to sigma-theta
-        #   Density contours formatting
-        cs = plt.contour(si, ti, dens, linestyles="dashed", colors="k")
-        plt.clabel(cs, fontsize=12, inline=1, fmt="%0.1f")
-
-    return _save_fig(ax, f_out) #   Save the figure if f_out is given else return the scatter axis
+# pending removal
